# Using C++23 expected for returns