--- a/README.md
+++ b/README.md
@@ -20,7 +20,8 @@ Specification for URL parser can be found from the
 - Parse and validate a URL
 
 ```cpp
-ada::url url = ada::parse("https://www.google.com");
+ada::result url = ada::parse("https://www.google.com");
+if(result) { /* URL is valid */ }
 // url.is_valid will return true
 // You are responsible to ensure that url.is_valid is true.
 ```
@@ -28,60 +29,60 @@ ada::url url = ada::parse("https://www.google.com");
 - Get/Update credentials
 
 ```cpp
-ada::url url = ada::parse("https://www.google.com");
-url.set_username("username");
-url.set_password("password");
-// ada.get_href() will return "https://username:password@www.google.com"
+ada::result url = ada::parse("https://www.google.com");
+url->set_username("username");
+url->set_password("password");
+// ada->get_href() will return "https://username:password@www.google.com"
 ```
 
 - Get/Update Protocol
 
 ```cpp
-ada::url url = ada::parse("https://www.google.com");
-url.set_protocol("wss");
-// url.get_protocol() will return "wss"
-// url.get_href() will return "wss://www.google.com"
+ada::result url = ada::parse("https://www.google.com");
+url->set_protocol("wss");
+// url->get_protocol() will return "wss"
+// url->get_href() will return "wss://www.google.com"
 ```
 
 - Get/Update host
 
 ```cpp
-ada::url url = ada::parse("https://www.google.com");
-url.set_host("github.com");
-// url.get_host() will return "github.com"
+ada::result url = ada::parse("https://www.google.com");
+url->set_host("github.com");
+// url->get_host() will return "github.com"
 // you can use `url.set_hostname` depending on your usage.
 ```
 
 - Get/Update port
 
 ```cpp
-ada::url url = ada::parse("https://www.google.com");
-url.set_port("8080");
-// url.get_port() will return "8080"
+ada::result url = ada::parse("https://www.google.com");
+url->set_port("8080");
+// url->get_port() will return "8080"
 ```
 
 - Get/Update pathname
 
 ```cpp
-ada::url url = ada::parse("https://www.google.com");
-url.set_pathname("/my-super-long-path")
-// url.get_pathname() will return "/my-super-long-path"
+ada::result url = ada::parse("https://www.google.com");
+url->set_pathname("/my-super-long-path")
+// url->get_pathname() will return "/my-super-long-path"
 ```
 
 - Get/Update search/query
 
 ```cpp
-ada::url url = ada::parse("https://www.google.com");
-url.set_search("target=self");
-// url.get_search() will return "?target=self"
+ada::result url = ada::parse("https://www.google.com");
+url->set_search("target=self");
+// url->get_search() will return "?target=self"
 ```
 
 - Get/Update hash/fragment
 
 ```cpp
-ada::url url = ada::parse("https://www.google.com");
-url.set_hash("is-this-the-real-life");
-// url.get_hash() will return "#is-this-the-real-life"
+ada::result url = ada::parse("https://www.google.com");
+url->set_hash("is-this-the-real-life");
+// url->get_hash() will return "#is-this-the-real-life"
 ```