--- a/include/ada/implementation.h
+++ b/include/ada/implementation.h
@@ -15,6 +15,9 @@
 #include "ada/state.h"
 
 namespace ada {
+  enum class errors {
+    generic_error
+  };
 
   /**
    * The URL parser takes a scalar value string input, with an optional null or base URL base (default null)
@@ -23,7 +26,7 @@ namespace ada {
    * @param input the string input to analyze.
    * @param base_url the optional string input to use as a base url.
    * @param encoding encoding (default to UTF-8)
-   * @return a parsed URL.  The caller *must* check that url.is_valid is true before using the url instance.
+   * @return a parsed URL.
    *
    * @example
    *
@@ -31,7 +34,7 @@ namespace ada {
    * auto url = ada::url parse("https://www.google.com");
    * ```
    */
-  ada_warn_unused ada::url parse(std::string_view input,
+  ada_warn_unused tl::expected<ada::url,ada::errors> parse(std::string_view input,
                                  const ada::url* base_url = nullptr,
                                  ada::encoding_type encoding = ada::encoding_type::UTF8);