--- a/src/implementation.cpp
+++ b/src/implementation.cpp
@@ -7,13 +7,15 @@
 
 namespace ada {
 
-  ada_warn_unused url parse(std::string_view input,
+  ada_warn_unused tl::expected<ada::url,ada::errors> parse(std::string_view input,
                             const ada::url* base_url,
                             ada::encoding_type encoding) {
     if(encoding != encoding_type::UTF8) {
       // @todo Add support for non UTF8 input
     }
-    return ada::parser::parse_url(input, base_url, encoding);
+    ada::url u = ada::parser::parse_url(input, base_url, encoding);
+    if(u.is_valid) { return tl::unexpected(errors::generic_error); }
+    return u;
   }
 
   ada_warn_unused std::string to_string(ada::encoding_type type) {