--- a/tests/installation/CMakeLists.txt
+++ b/tests/installation/CMakeLists.txt
@@ -45,10 +45,10 @@ file(WRITE main.cpp "
 #include <iostream>
 
 int main(int , char *[]) {
-  ada::url url = ada::parse(\"https://www.google.com\");
-  url.set_protocol(\"http\");
-  std::cout << url.get_protocol() << std::endl;
-  std::cout << url.get_host() << std::endl;
+  ada::result url = ada::parse(\"https://www.google.com\");
+  url->set_protocol(\"http\");
+  std::cout << url->get_protocol() << std::endl;
+  std::cout << url->get_host() << std::endl;
   return EXIT_SUCCESS;
 }")