This is a demo/prototype of how we might use `std::expected` (a C++23 feature) in ada. We use a library because it is not yet supported in mainstream systems (since most compilers do not support C++23 yet).

Fixes https://github.com/ada-url/ada/issues/151

The binary bloat is small (700 extra bytes more or less).

The performance effect is nil on bench:

Before...

```
BasicBench_AdaURL_With_Move       2299 ns         2299 ns       305794 speed=351.47M/s time/byte=2.84519ns time/url=208.992ns url/s=4.78486M/s
```

After:

```
BasicBench_AdaURL_With_Move       2298 ns         2298 ns       305233 speed=351.673M/s time/byte=2.84355ns time/url=208.872ns url/s=4.78763M/s
```


It is all a matter of deciding whether we like the API.

Todo:

- [ ] We don't actually need the is_valid flag, it should be removed.
- [ ] We could use more than just the generic error.
- [X] We should make sure that we understand the performance effect of this PR before merging it, if we merge it.

Note that the expected code is public domain. So we are good.