@lemire 2023-02-03
@anonrig  This is ready for review. It could be merged.

It will impact your work on Node, but it should be only superficial (some copy-paste).